applications/model_trainer/pyproject.toml
[tool.poetry]
name = "model_trainer"
version = "0.1.0"
description = ""
authors = ["gatheluck <gatheluck+tech@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.9.0"
torchvision = "^0.10.0"
hydra-core = "^1.1.0"
pytorch-lightning = "^1.4.0"
comet-ml = "^3.13.2"
mlflow = "^1.19.0"
albumentations = "^1.0.3"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.7b0"
isort = "^5.9.2"
flake8 = "^3.9.2"
pytest-cov = "^2.12.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
